requirements.txt
pandas~=1.3.4
numpy~=1.20.3
future~=0.18.2
scikit-learn~=0.24.2
torchvision~=0.11.3
networkx~=2.6.3
matplotlib~=3.4.3
tqdm~=4.62.3
cvxopt~=1.3.0
cvxpy~=1.3.0